tech: fix code owner and create action to update node version (#325)
This pull request introduces automation for keeping the project's
Node.js version up to date and makes a small correction to the code
owners file. The most significant change is the addition of a GitHub
Actions workflow that updates the `.nvmrc` file with the latest Node.js
version every week and creates a pull request for the update.

Automation and workflow improvements:

* Added a new GitHub Actions workflow
(`.github/workflows/update-nvmrc.yml`) that runs weekly to fetch the
latest Node.js version, update the `.nvmrc` file, and create a pull
request with the changes.

Maintenance and configuration:

* Fixed the syntax in the `.github/CODEOWNERS` file by prefixing the
username with `@` to properly assign code ownership.