feat: add the setting diagnostic_mode (#378)
This pull request introduces a new "diagnostic mode" setting to the
application, making it available in the user interface and providing
translations for English, Spanish, French, and Italian. The main changes
are the addition of the new setting and its localization.

Settings feature update:

* Added a new `diagnostic_mode` checkbox setting to `src/settings.json`,
with a default value of false.

Localization updates:

* Added the `diagnostic_mode` translation string to `locales/en.json`
for English.
* Added the `diagnostic_mode` translation string to `locales/es.json`
for Spanish.
* Added the `diagnostic_mode` translation string to `locales/fr.json`
for French.
* Added the `diagnostic_mode` translation string to `locales/it.json`
for Italian.